Refactoring du code
