+++ admission/tests/tests.py
@@ -142,138 +142,3 @@ class SendMailTest(TestCase):
                                                          activation_code,
                                                          an_applicant,
                                                          'account_activation'))
-
-
-class AssimilationCriteriaTest(TestCase):
-
-    def test_criteria1(self):
-        list_document_type = []
-        list_document_type = assimilation_criteria.criteria1(list_document_type)
-        self.assertTrue(len(list_document_type) == 2)
-
-        list_document_type = assimilation_criteria.criteria1(None)
-        self.assertTrue(len(list_document_type) == 2)
-
-        list_document_type = []
-        assimilation_doc = assimilation_criteria.AssimilationDoc()
-        assimilation_doc.criteria_id = 1
-        list_document_type.append(assimilation_doc)
-        list_document_type = assimilation_criteria.criteria1(list_document_type)
-        self.assertTrue(len(list_document_type) == 3)
-
-    def test_criteria2(self):
-        list_document_type = []
-        list_document_type = assimilation_criteria.criteria2(list_document_type)
-        self.assertTrue(len(list_document_type) == 5)
-
-        list_document_type = assimilation_criteria.criteria2(None)
-        self.assertTrue(len(list_document_type) == 5)
-
-        list_document_type = []
-        assimilation_doc = assimilation_criteria.AssimilationDoc()
-        assimilation_doc.criteria_id = 2
-        list_document_type.append(assimilation_doc)
-        list_document_type = assimilation_criteria.criteria2(list_document_type)
-        self.assertTrue(len(list_document_type) == 6)
-
-    def test_find_list_document_type_by_criteria(self):
-        list_document_type = assimilation_criteria.find_list_document_type_by_criteria(None)
-        self.assertTrue(len(list_document_type) == 0)
-
-        list_document_type = assimilation_criteria.find_list_document_type_by_criteria("1")
-        self.assertTrue(len(list_document_type) == 0)
-
-        list_document_type = assimilation_criteria.find_list_document_type_by_criteria(1)
-        self.assertTrue(len(list_document_type) == 2)
-
-    def test_get_list_docs(self):
-        list_document_type = assimilation_criteria.get_list_docs(None)
-        self.assertTrue(len(list_document_type) == 0)
-
-        list_document_type = assimilation_criteria.get_list_docs("1")
-        self.assertTrue(len(list_document_type) == 0)
-
-        list_document_type = assimilation_criteria.get_list_docs(3)
-        self.assertTrue(len(list_document_type) == 9)
-
-
-class SecondaryEducationTest(TestCase):
-
-    def setUp(self):
-        self.factory = RequestFactory()
-        self.user = User.objects.create_user(
-            username='jacob', email='jacob@gmail.com', password='top_secret')
-        self.applicant = data_model.create_applicant_by_user(self.user)
-
-    def test_get_secondary_education_exams_data(self):
-        secondary_education_record = None
-        list_secondary_education_exams = secondary_education.\
-            get_secondary_education_exams_data(secondary_education_record)
-        self.assertTrue(len(list_secondary_education_exams) == 0)
-
-        secondary_education_record = data_model.create_secondary_education_with_exams()
-        list_secondary_education_exams = secondary_education.\
-            get_secondary_education_exams_data(secondary_education_record)
-        self.assertTrue(len(list_secondary_education_exams) == 3)
-
-    def test_is_local_language_exam_needed(self):
-        self.assertFalse(secondary_education.is_local_language_exam_needed(None))
-
-        # a_user = data_model.create_user()
-        self.assertFalse(secondary_education.is_local_language_exam_needed(self.user))
-
-        an_application = data_model.create_application(self.applicant)
-        self.assertFalse(secondary_education.is_local_language_exam_needed(self.user))
-
-        an_application.offer_year.grade_type = data_model.create_grade_type('BACHELOR')
-        an_application.offer_year.save()
-        self.assertTrue(secondary_education.is_local_language_exam_needed(self.user))
-
-        an_application.offer_year.grade_type = data_model.create_grade_type('BACHELORZ')
-        an_application.offer_year.save()
-        self.assertFalse(secondary_education.is_local_language_exam_needed(self.user))
-
-    def test_secondary_education_exam_update(self):
-        secondary_education_record = data_model.create_secondary_education_with_exams()
-        type = 'ADMISSION'
-        secondary_education_exam = data_model.create_secondary_education_exam(secondary_education_record, type)
-
-        try:
-            secondary_education.secondary_education_exam_update(secondary_education_record,
-                                                                type,
-                                                                secondary_education_exam)
-        except ExceptionType:
-            self.fail("secondary_education_exam_update raised ExceptionType unexpectedly!")
-
-        try:
-            secondary_education.secondary_education_exam_update(None, type, secondary_education_exam)
-        except ExceptionType:
-            self.fail("secondary_education_exam_update raised ExceptionType unexpectedly!")
-
-        try:
-            secondary_education.secondary_education_exam_update(None, None, secondary_education_exam)
-        except ExceptionType:
-            self.fail("secondary_education_exam_update raised ExceptionType unexpectedly!")
-
-        try:
-            secondary_education.secondary_education_exam_update(None, None, None)
-        except Exception:
-            self.fail("secondary_education_exam_update raised ExceptionType unexpectedly!")
-
-    def test_get_secondary_education_files_data(self):
-        try:
-            secondary_education.get_secondary_education_files_data(None)
-        except:
-            self.fail("get_secondary_education_files_data raised ExceptionType unexpectedly!")
-        an_application = data_model.create_application(self.applicant)
-        try:
-            secondary_education.get_secondary_education_files_data(an_application)
-        except Exception:
-            self.fail("get_secondary_education_files_data raised ExceptionType unexpectedly!")
-
-        an_application_document_file = data_model.create_application_document_file(an_application,
-                                                                                   self.user,
-                                                                                   'NATIONAL_DIPLOMA_VERSO')
-
-        dict = secondary_education.get_secondary_education_files_data(an_application)
-        self.assertTrue(dict['national_diploma_verso'] == an_application_document_file)
