+++ admission/views/demande_validation.py
@@ -93,7 +93,7 @@ def validate_profil(applicant, user):
                 return False
     assimilation_criteria_list = mdl.applicant_assimilation_criteria.find_by_applicant(applicant)
     for assimilation_criteria in assimilation_criteria_list:
-        docs_needed = assimilation_criteria_view.get_list_docs(assimilation_criteria.criteria.id)
+        docs_needed = assimilation_criteria_view.get_list_documents_descriptions(assimilation_criteria.criteria.id)
         for doc_needed in docs_needed:
             doc = mdl_common.document_file.search(user, doc_needed)
             if not doc.exists():
